pinecone-embedding/pyproject.toml
[project]
name = "rag-ingest"
version = "0.1.0"
description = "Pinecone ingest and embed script"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3==1.40.69",
    "numpy==2.2",
    "pinecone==7.3.0",
    "polars==1.25.2",
    "python-dotenv>=1.2.1",
    "requests==2.32.4",
    "tqdm>=4.67.1",
]
